refactor: 선발 절차 과정의 강의를 수강 신청

+++ src/main/java/nextstep/courses/domain/repository/CourseRepository.java
@@ -3,7 +3,7 @@ package nextstep.courses.domain.repository;
 import nextstep.courses.domain.model.Course;
 
 public interface CourseRepository {
-    int save(Course course);
+    long save(Course course);
 
     Course findById(Long id);
 }
