+++ src/main/java/nextstep/courses/domain/repository/SessionRepository.java
@@ -3,7 +3,7 @@ package nextstep.courses.domain.repository;
 import nextstep.courses.domain.model.Session;
 
 public interface SessionRepository {
-    int save(Session session);
+    long save(Session session);
 
     Session findById(Long id);
 }
