+++ src/main/java/nextstep/courses/domain/repository/StudentRepository.java
@@ -3,7 +3,7 @@ package nextstep.courses.domain.repository;
 import nextstep.courses.domain.model.Student;
 
 public interface StudentRepository {
-    int save(Student course);
+    long save(Student course);
 
     Student findById(Long id);
 }
